third_party/blink/web_tests/external/wpt/css/css-conditional/container-queries/scroll-state/multiple-scroll-state-containers-comma-separated-queries-expected.txt

This is a testharness.js-based test.
[FAIL] Should match the named outer container for scroll-state(scrollable)
  assert_equals: expected "yes" but got "no"
[FAIL] Match the #combined outer container which is also a size container
  assert_equals: expected "yes" but got "no"
Harness: the test ran to completion.
